requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
tiktoken
pydantic
python-dotenv
streamlit
requests
ragas
apache-airflow
